"Fixed" an infuriating text alignment issue.
Not all SVG viewers will respect the dominant-baseline tag for text vertical
alignment.  Firefox does, as does Chrome, but Inkscape does not.  Inkscape
can produce nice PDFs, but Firefox won't allow page size changes.  And
pandoc is oblivious to SVGs so we have to PDF graphs first.

How ridiculous!

Right now, manually guessed that 0.35 of a font's height is equal to the distance
from the normal baseline to the  midpoint.  Not sure that's true.  Probably not.
But at least it looks kind of right.