Added `.cursor` to files and byte arrays for line-by-line reading.
This gives you access to a `kse.eio.cursor.ByteCursor` which gives
you a variety of handy methods for traversing the contents of a file,
which it does with caching to make it reasonably efficient.

The key methods are `to`, `hasNext` and `next` (to get more lines),
and `preparePrev` (which both prepares to read the previous line and says
whether one exists, like `hasNext` except with changes to the read location)
and `prevOrNull` (which gives `null` if the previous string does not
exist--or you can use `prevOption` if you don't mind wrapping every string).